Deep Learning/PyTorch Lightning/pyproject.toml
[project]
name = "pytorch-lightning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "numpy<2",
    "omegaconf>=2.3.0",
    "rich>=14.3.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch==2.2.1",
    "torch-tb-profiler>=0.4.3",
    "torchmetrics>=1.0.0",
    "torchvision==0.17.1",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu118"
default = true
